refactor: make use of db types and common error where possible
